# virsh-sandbox/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing virtual machine sandboxes using libvirt",
        "title": "virsh-sandbox API",
        "contact": {},
        "version": "0.0.1-beta"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/v1/ansible/jobs": {
            "post": {
                "description": "Creates a new Ansible playbook execution job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ansible"
                ],
                "summary": "Create Ansible job",
                "parameters": [
                    {
                        "description": "Job creation parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_ansible.JobRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_ansible.JobResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_error.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/ansible/jobs/{job_id}": {
            "get": {
                "description": "Gets the status of an Ansible job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ansible"
                ],
                "summary": "Get Ansible job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_ansible.Job"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/virsh-sandbox_internal_error.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/ansible/jobs/{job_id}/stream": {
            "get": {
                "description": "Connects via WebSocket to run an Ansible job and stream output",
                "tags": [
                    "Ansible"
                ],
                "summary": "Stream Ansible job output",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols - WebSocket connection established",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Invalid job ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Job already started or finished",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_ansible.Job": {
            "type": "object",
            "properties": {
                "check": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "playbook": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/internal_ansible.JobStatus"
                },
                "vm_name": {
                    "type": "string"
                }
            }
        },
        "internal_ansible.JobRequest": {
            "type": "object",
            "properties": {
                "check": {
                    "type": "boolean"
                },
                "playbook": {
                    "type": "string"
                },
                "vm_name": {
                    "type": "string"
                }
            }
        },
        "internal_ansible.JobResponse": {
            "type": "object",
            "properties": {
                "job_id": {
                    "type": "string"
                },
                "ws_url": {
                    "type": "string"
                }
            }
        },
        "internal_ansible.JobStatus": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "finished",
                "failed"
            ],
            "x-enum-varnames": [
                "JobStatusPending",
                "JobStatusRunning",
                "JobStatusFinished",
                "JobStatusFailed"
            ]
        },
...
        "virsh-sandbox_internal_error.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
...
}